+++ c/csv/freadLookups.h
